Remove unsound `TxConnection` API
You can move out a `TxConnection` with `mem::swap`. That must've been
obvious. Why on earth did I overlook it?

`TxConnection` was meant to be a hack to replicate the API of Diesel 2,
which will use mutable references for `Connection` methods. Now that
this approach failed, however, I cannot come up with a way to achieve
the same purpose w/o GAT or runtime checks (I'd bet such a way doesn't
exist), so we are going to abandon interoperability between Diesel 1
and 2 for now.